Rewrite modem detection code
Instead of listing the available tty devices and finding out the other
info from there, start with listing all usb devices and find out if each
of them has tty devices. This should be a bit more robust and allows for
listing all available devices in the future (instead of always stopping
at the first match), without risk of returning a device multiple times
(once for each tty it offers).