Store the control and data tty explicitely
Instead of storing the name of the first tty and deriving the control
and data tty names from the first tty and their index using fragile
string manipulation, now we just store copy tty names explicitely from
sysfs.

Note that if the tty devices are not numbered in the same order as the
usb subdevices / endpoints, this will change the tty devices used, but
we assume the kernel assigns the tty devices to the endpoints in
order, so this should work out.