Allow listing devices without any profile
Previously, any device with no profile, or an invalid profile, was not
listed or used. Now, these devices are not used for connecting, but they
are listed with -l. This allows a GUI to display these devices and allow
the user to create a custom configuration profile for these devices to
get them working.