Add Huawei modem data from Huawei Linux driver
This adds the udev rules file from the Huawei Linux driver and a python
script to convert to a proper header file that udiald can include.

This makes sure that for all the listed Huawei devices, at least the
data and control port are already correctly set, making it likely that a
lot of these devices will work out of the box now.