Rewrite reply parsing
Instead of piling the entire reply in a single buffer, it is now read
and saved line by line and it stops reading as soon as the end of the
reply is found.

While reading, it also (optionally) looks for a specific line prefix and
stores a pointer to the first matching line as "result_line".

This change should make reply parsing a bit more reliable, but also
ensures that any raw errors logged don't have embedded newlines, which
makes them work as uci values.