Store the failed pin in a separate uci section
Before, it was stored in the "wan" (or whatever was configured) section
of the network uci state, but (at least on the Fonera) this section is
cleared out on every connection attempt, which voided the point of
storing the failed pin. Now, a separate section in the network config
file is used to store the failed pin.